♻️ refactor: isolate web search providers
